backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.3

# Security / Auth
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4

# AI
anthropic==0.39.0
openai==1.54.3  # Optional: for OpenAI embeddings as alternative

# Data processing
pandas==2.2.3
numpy==1.26.4
scipy==1.14.0  # For Pearson correlation

# Visualization
plotly==5.24.1
kaleido==0.2.1  # For Plotly PNG export

# Email
aiosmtplib==3.0.1
jinja2==3.1.4  # Email templates

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12  # File uploads

# PDF export
reportlab==4.2.5

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0  # Retry logic

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  # TestClient
